refactor: rename store components
