+++ src/entities/storeList/ui/index.ts
@@ -1,4 +1,5 @@
-export * from "./SelectedStoreCard";
-export * from "./StoreCard";
+export * from "./SelectedStorePreviewCard";
 export * from "./StoreLevelList";
-export * from "./StoreList";
+export * from "./StoreListSection";
+export * from "./StorePreviewCard";
+export * from "./StorePreviewList";
